Refactor httpApi into two parts
