+++ src/Microsoft.AspNetCore.Server.HttpSys/NativeInterop/CookedUrl.cs
@@ -10,9 +10,9 @@ namespace Microsoft.AspNetCore.Server.HttpSys
     // Note this type should only be used while the request buffer remains pinned
     internal class CookedUrl
     {
-        private readonly HttpApi.HTTP_COOKED_URL _nativeCookedUrl;
+        private readonly HttpNativeStructs.HTTP_COOKED_URL _nativeCookedUrl;
 
-        internal CookedUrl(HttpApi.HTTP_COOKED_URL nativeCookedUrl)
+        internal CookedUrl(HttpNativeStructs.HTTP_COOKED_URL nativeCookedUrl)
         {
             _nativeCookedUrl = nativeCookedUrl;
         }
