+++ src/Microsoft.AspNetCore.Server.HttpSys/NativeInterop/HttpSysResponseHeader.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Server.HttpSys
+namespace Microsoft.AspNetCore.HttpSys.Internal
 {
     internal enum HttpSysResponseHeader
     {
