+++ src/Microsoft.AspNetCore.Server.HttpSys/RequestProcessing/Response.cs
@@ -25,7 +25,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         private TimeSpan? _cacheTtl;
         private long _expectedBodyLength;
         private BoundaryType _boundaryType;
-        private HttpApi.HTTP_RESPONSE_V2 _nativeResponse;
+        private HttpNativeStructs.HTTP_RESPONSE_V2 _nativeResponse;
 
         internal Response(RequestContext requestContext)
         {
@@ -34,7 +34,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
             Headers = new HeaderCollection();
             // We haven't started yet, or we're just buffered, we can clear any data, headers, and state so
             // that we can start over (e.g. to write an error message).
-            _nativeResponse = new HttpApi.HTTP_RESPONSE_V2();
+            _nativeResponse = new HttpNativeStructs.HTTP_RESPONSE_V2();
             Headers.IsReadOnly = false;
             Headers.Clear();
             _reasonPhrase = null;
@@ -250,9 +250,9 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         // What would we loose by bypassing HttpSendHttpResponse?
         //
         // TODO: Consider using the HTTP_SEND_RESPONSE_FLAG_BUFFER_DATA flag for most/all responses rather than just Opaque.
-        internal unsafe uint SendHeaders(HttpApi.HTTP_DATA_CHUNK[] dataChunks,
+        internal unsafe uint SendHeaders(HttpNativeStructs.HTTP_DATA_CHUNK[] dataChunks,
             ResponseStreamAsyncResult asyncResult,
-            HttpApi.HTTP_FLAGS flags,
+            HttpNativeStructs.HTTP_FLAGS flags,
             bool isOpaqueUpgrade)
         {
             Debug.Assert(!HasStarted, "HttpListenerResponse::SendHeaders()|SentHeaders is true.");
@@ -274,7 +274,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                     var handle = GCHandle.Alloc(dataChunks, GCHandleType.Pinned);
                     pinnedHeaders.Add(handle);
                     _nativeResponse.Response_V1.EntityChunkCount = (ushort)dataChunks.Length;
-                    _nativeResponse.Response_V1.pEntityChunks = (HttpApi.HTTP_DATA_CHUNK*)handle.AddrOfPinnedObject();
+                    _nativeResponse.Response_V1.pEntityChunks = (HttpNativeStructs.HTTP_DATA_CHUNK*)handle.AddrOfPinnedObject();
                 }
                 else if (asyncResult != null && asyncResult.DataChunks != null)
                 {
@@ -287,10 +287,10 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                     _nativeResponse.Response_V1.pEntityChunks = null;
                 }
 
-                var cachePolicy = new HttpApi.HTTP_CACHE_POLICY();
+                var cachePolicy = new HttpNativeStructs.HTTP_CACHE_POLICY();
                 if (_cacheTtl.HasValue && _cacheTtl.Value > TimeSpan.Zero)
                 {
-                    cachePolicy.Policy = HttpApi.HTTP_CACHE_POLICY_TYPE.HttpCachePolicyTimeToLive;
+                    cachePolicy.Policy = HttpNativeStructs.HTTP_CACHE_POLICY_TYPE.HttpCachePolicyTimeToLive;
                     cachePolicy.SecondsToLive = (uint)Math.Min(_cacheTtl.Value.Ticks / TimeSpan.TicksPerSecond, Int32.MaxValue);
                 }
 
@@ -299,7 +299,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                 {
                     _nativeResponse.Response_V1.ReasonLength = (ushort)reasonPhraseBytes.Length;
                     _nativeResponse.Response_V1.pReason = (sbyte*)pReasonPhrase;
-                    fixed (HttpApi.HTTP_RESPONSE_V2* pResponse = &_nativeResponse)
+                    fixed (HttpNativeStructs.HTTP_RESPONSE_V2* pResponse = &_nativeResponse)
                     {
                         statusCode =
                             HttpApi.HttpSendHttpResponse(
@@ -331,14 +331,14 @@ namespace Microsoft.AspNetCore.Server.HttpSys
             return statusCode;
         }
 
-        internal HttpApi.HTTP_FLAGS ComputeHeaders(long writeCount, bool endOfRequest = false)
+        internal HttpNativeStructs.HTTP_FLAGS ComputeHeaders(long writeCount, bool endOfRequest = false)
         {
             if (StatusCode == (ushort)StatusCodes.Status401Unauthorized)
             {
                 RequestContext.Server.Options.Authentication.SetAuthenticationChallenge(RequestContext);
             }
 
-            var flags = HttpApi.HTTP_FLAGS.NONE;
+            var flags = HttpNativeStructs.HTTP_FLAGS.NONE;
             Debug.Assert(!HasComputedHeaders, nameof(HasComputedHeaders) + " is true.");
             _responseState = ResponseState.ComputedHeaders;
 
@@ -415,7 +415,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                 {
                     Headers.Append(HttpKnownHeaderNames.Connection, Constants.Close);
                 }
-                flags = HttpApi.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_DISCONNECT;
+                flags = HttpNativeStructs.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_DISCONNECT;
             }
 
             return flags;
@@ -429,8 +429,8 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         private unsafe List<GCHandle> SerializeHeaders(bool isOpaqueUpgrade)
         {
             Headers.IsReadOnly = true; // Prohibit further modifications.
-            HttpApi.HTTP_UNKNOWN_HEADER[] unknownHeaders = null;
-            HttpApi.HTTP_RESPONSE_INFO[] knownHeaderInfo = null;
+            HttpNativeStructs.HTTP_UNKNOWN_HEADER[] unknownHeaders = null;
+            HttpNativeStructs.HTTP_RESPONSE_INFO[] knownHeaderInfo = null;
             List<GCHandle> pinnedHeaders;
             GCHandle gcHandle;
 
@@ -453,11 +453,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                     continue;
                 }
                 // See if this is an unknown header
-                lookup = HttpApi.HTTP_RESPONSE_HEADER_ID.IndexOfKnownHeader(headerPair.Key);
+                lookup = HttpNativeStructs.HTTP_RESPONSE_HEADER_ID.IndexOfKnownHeader(headerPair.Key);
 
                 // Http.Sys doesn't let us send the Connection: Upgrade header as a Known header.
                 if (lookup == -1 ||
-                    (isOpaqueUpgrade && lookup == (int)HttpApi.HTTP_RESPONSE_HEADER_ID.Enum.HttpHeaderConnection))
+                    (isOpaqueUpgrade && lookup == (int)HttpNativeStructs.HTTP_RESPONSE_HEADER_ID.Enum.HttpHeaderConnection))
                 {
                     numUnknownHeaders += headerPair.Value.Count;
                 }
@@ -470,7 +470,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
 
             try
             {
-                fixed (HttpApi.HTTP_KNOWN_HEADER* pKnownHeaders = &_nativeResponse.Response_V1.Headers.KnownHeaders)
+                fixed (HttpNativeStructs.HTTP_KNOWN_HEADER* pKnownHeaders = &_nativeResponse.Response_V1.Headers.KnownHeaders)
                 {
                     foreach (var headerPair in Headers)
                     {
@@ -480,18 +480,18 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                         }
                         headerName = headerPair.Key;
                         StringValues headerValues = headerPair.Value;
-                        lookup = HttpApi.HTTP_RESPONSE_HEADER_ID.IndexOfKnownHeader(headerName);
+                        lookup = HttpNativeStructs.HTTP_RESPONSE_HEADER_ID.IndexOfKnownHeader(headerName);
 
                         // Http.Sys doesn't let us send the Connection: Upgrade header as a Known header.
                         if (lookup == -1 ||
-                            (isOpaqueUpgrade && lookup == (int)HttpApi.HTTP_RESPONSE_HEADER_ID.Enum.HttpHeaderConnection))
+                            (isOpaqueUpgrade && lookup == (int)HttpNativeStructs.HTTP_RESPONSE_HEADER_ID.Enum.HttpHeaderConnection))
                         {
                             if (unknownHeaders == null)
                             {
-                                unknownHeaders = new HttpApi.HTTP_UNKNOWN_HEADER[numUnknownHeaders];
+                                unknownHeaders = new HttpNativeStructs.HTTP_UNKNOWN_HEADER[numUnknownHeaders];
                                 gcHandle = GCHandle.Alloc(unknownHeaders, GCHandleType.Pinned);
                                 pinnedHeaders.Add(gcHandle);
-                                _nativeResponse.Response_V1.Headers.pUnknownHeaders = (HttpApi.HTTP_UNKNOWN_HEADER*)gcHandle.AddrOfPinnedObject();
+                                _nativeResponse.Response_V1.Headers.pUnknownHeaders = (HttpNativeStructs.HTTP_UNKNOWN_HEADER*)gcHandle.AddrOfPinnedObject();
                             }
 
                             for (int headerValueIndex = 0; headerValueIndex < headerValues.Count; headerValueIndex++)
@@ -526,24 +526,24 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                         {
                             if (knownHeaderInfo == null)
                             {
-                                knownHeaderInfo = new HttpApi.HTTP_RESPONSE_INFO[numKnownMultiHeaders];
+                                knownHeaderInfo = new HttpNativeStructs.HTTP_RESPONSE_INFO[numKnownMultiHeaders];
                                 gcHandle = GCHandle.Alloc(knownHeaderInfo, GCHandleType.Pinned);
                                 pinnedHeaders.Add(gcHandle);
-                                _nativeResponse.pResponseInfo = (HttpApi.HTTP_RESPONSE_INFO*)gcHandle.AddrOfPinnedObject();
+                                _nativeResponse.pResponseInfo = (HttpNativeStructs.HTTP_RESPONSE_INFO*)gcHandle.AddrOfPinnedObject();
                             }
 
-                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].Type = HttpApi.HTTP_RESPONSE_INFO_TYPE.HttpResponseInfoTypeMultipleKnownHeaders;
-                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].Length = (uint)Marshal.SizeOf<HttpApi.HTTP_MULTIPLE_KNOWN_HEADERS>();
+                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].Type = HttpNativeStructs.HTTP_RESPONSE_INFO_TYPE.HttpResponseInfoTypeMultipleKnownHeaders;
+                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].Length = (uint)Marshal.SizeOf<HttpNativeStructs.HTTP_MULTIPLE_KNOWN_HEADERS>();
 
-                            HttpApi.HTTP_MULTIPLE_KNOWN_HEADERS header = new HttpApi.HTTP_MULTIPLE_KNOWN_HEADERS();
+                            HttpNativeStructs.HTTP_MULTIPLE_KNOWN_HEADERS header = new HttpNativeStructs.HTTP_MULTIPLE_KNOWN_HEADERS();
 
-                            header.HeaderId = (HttpApi.HTTP_RESPONSE_HEADER_ID.Enum)lookup;
-                            header.Flags = HttpApi.HTTP_RESPONSE_INFO_FLAGS.PreserveOrder; // TODO: The docs say this is for www-auth only.
+                            header.HeaderId = (HttpNativeStructs.HTTP_RESPONSE_HEADER_ID.Enum)lookup;
+                            header.Flags = HttpNativeStructs.HTTP_RESPONSE_INFO_FLAGS.PreserveOrder; // TODO: The docs say this is for www-auth only.
 
-                            HttpApi.HTTP_KNOWN_HEADER[] nativeHeaderValues = new HttpApi.HTTP_KNOWN_HEADER[headerValues.Count];
+                            HttpNativeStructs.HTTP_KNOWN_HEADER[] nativeHeaderValues = new HttpNativeStructs.HTTP_KNOWN_HEADER[headerValues.Count];
                             gcHandle = GCHandle.Alloc(nativeHeaderValues, GCHandleType.Pinned);
                             pinnedHeaders.Add(gcHandle);
-                            header.KnownHeaders = (HttpApi.HTTP_KNOWN_HEADER*)gcHandle.AddrOfPinnedObject();
+                            header.KnownHeaders = (HttpNativeStructs.HTTP_KNOWN_HEADER*)gcHandle.AddrOfPinnedObject();
 
                             for (int headerValueIndex = 0; headerValueIndex < headerValues.Count; headerValueIndex++)
                             {
@@ -560,7 +560,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                             // This type is a struct, not an object, so pinning it causes a boxed copy to be created. We can't do that until after all the fields are set.
                             gcHandle = GCHandle.Alloc(header, GCHandleType.Pinned);
                             pinnedHeaders.Add(gcHandle);
-                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].pInfo = (HttpApi.HTTP_MULTIPLE_KNOWN_HEADERS*)gcHandle.AddrOfPinnedObject();
+                            knownHeaderInfo[_nativeResponse.ResponseInfoCount].pInfo = (HttpNativeStructs.HTTP_MULTIPLE_KNOWN_HEADERS*)gcHandle.AddrOfPinnedObject();
 
                             _nativeResponse.ResponseInfoCount++;
                         }
@@ -596,9 +596,9 @@ namespace Microsoft.AspNetCore.Server.HttpSys
 
             // TODO: Send headers async?
             ulong errorCode = SendHeaders(null, null,
-                HttpApi.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_OPAQUE |
-                HttpApi.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_MORE_DATA |
-                HttpApi.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_BUFFER_DATA,
+                HttpNativeStructs.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_OPAQUE |
+                HttpNativeStructs.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_MORE_DATA |
+                HttpNativeStructs.HTTP_FLAGS.HTTP_SEND_RESPONSE_FLAG_BUFFER_DATA,
                 true);
 
             if (errorCode != ErrorCodes.ERROR_SUCCESS)
