+++ src/Microsoft.AspNetCore.Server.HttpSys/TimeoutManager.cs
@@ -16,7 +16,7 @@ namespace Microsoft.AspNetCore.Server.HttpSys
     public sealed class TimeoutManager
     {
         private static readonly int TimeoutLimitSize =
-            Marshal.SizeOf<HttpApi.HTTP_TIMEOUT_LIMIT_INFO>();
+            Marshal.SizeOf<HttpNativeStructs.HTTP_TIMEOUT_LIMIT_INFO>();
 
         private UrlGroup _urlGroup;
         private int[] _timeouts;
@@ -48,11 +48,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         {
             get
             {
-                return GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.EntityBody);
+                return GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.EntityBody);
             }
             set
             {
-                SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.EntityBody, value);
+                SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.EntityBody, value);
             }
         }
 
@@ -71,11 +71,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         {
             get
             {
-                return GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.DrainEntityBody);
+                return GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.DrainEntityBody);
             }
             set
             {
-                SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.DrainEntityBody, value);
+                SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.DrainEntityBody, value);
             }
         }
 
@@ -89,11 +89,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         {
             get
             {
-                return GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.RequestQueue);
+                return GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.RequestQueue);
             }
             set
             {
-                SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.RequestQueue, value);
+                SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.RequestQueue, value);
             }
         }
 
@@ -108,11 +108,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         {
             get
             {
-                return GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.IdleConnection);
+                return GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.IdleConnection);
             }
             set
             {
-                SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.IdleConnection, value);
+                SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.IdleConnection, value);
             }
         }
 
@@ -128,11 +128,11 @@ namespace Microsoft.AspNetCore.Server.HttpSys
         {
             get
             {
-                return GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.HeaderWait);
+                return GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.HeaderWait);
             }
             set
             {
-                SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE.HeaderWait, value);
+                SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE.HeaderWait, value);
             }
         }
 
@@ -168,13 +168,13 @@ namespace Microsoft.AspNetCore.Server.HttpSys
 
         #region Helpers
 
-        private TimeSpan GetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE type)
+        private TimeSpan GetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE type)
         {
             // Since we maintain local state, GET is local.
             return new TimeSpan(0, 0, (int)_timeouts[(int)type]);
         }
 
-        private void SetTimeSpanTimeout(HttpApi.HTTP_TIMEOUT_TYPE type, TimeSpan value)
+        private void SetTimeSpanTimeout(HttpNativeStructs.HTTP_TIMEOUT_TYPE type, TimeSpan value)
         {
             // All timeouts are defined as USHORT in native layer (except MinSendRate, which is ULONG). Make sure that
             // timeout value is within range.
@@ -208,25 +208,25 @@ namespace Microsoft.AspNetCore.Server.HttpSys
                 return;
             }
 
-            var timeoutinfo = new HttpApi.HTTP_TIMEOUT_LIMIT_INFO();
+            var timeoutinfo = new HttpNativeStructs.HTTP_TIMEOUT_LIMIT_INFO();
 
-            timeoutinfo.Flags = HttpApi.HTTP_FLAGS.HTTP_PROPERTY_FLAG_PRESENT;
+            timeoutinfo.Flags = HttpNativeStructs.HTTP_FLAGS.HTTP_PROPERTY_FLAG_PRESENT;
             timeoutinfo.DrainEntityBody =
-                (ushort)timeouts[(int)HttpApi.HTTP_TIMEOUT_TYPE.DrainEntityBody];
+                (ushort)timeouts[(int)HttpNativeStructs.HTTP_TIMEOUT_TYPE.DrainEntityBody];
             timeoutinfo.EntityBody =
-                (ushort)timeouts[(int)HttpApi.HTTP_TIMEOUT_TYPE.EntityBody];
+                (ushort)timeouts[(int)HttpNativeStructs.HTTP_TIMEOUT_TYPE.EntityBody];
             timeoutinfo.RequestQueue =
-                (ushort)timeouts[(int)HttpApi.HTTP_TIMEOUT_TYPE.RequestQueue];
+                (ushort)timeouts[(int)HttpNativeStructs.HTTP_TIMEOUT_TYPE.RequestQueue];
             timeoutinfo.IdleConnection =
-                (ushort)timeouts[(int)HttpApi.HTTP_TIMEOUT_TYPE.IdleConnection];
+                (ushort)timeouts[(int)HttpNativeStructs.HTTP_TIMEOUT_TYPE.IdleConnection];
             timeoutinfo.HeaderWait =
-                (ushort)timeouts[(int)HttpApi.HTTP_TIMEOUT_TYPE.HeaderWait];
+                (ushort)timeouts[(int)HttpNativeStructs.HTTP_TIMEOUT_TYPE.HeaderWait];
             timeoutinfo.MinSendRate = minSendBytesPerSecond;
 
             var infoptr = new IntPtr(&timeoutinfo);
 
             _urlGroup.SetProperty(
-                HttpApi.HTTP_SERVER_PROPERTY.HttpServerTimeoutsProperty,
+                HttpNativeStructs.HTTP_SERVER_PROPERTY.HttpServerTimeoutsProperty,
                 infoptr, (uint)TimeoutLimitSize);
         }
